Update README.md
Hey there!
Line 132 and 135 had excluded 'self_' part from the name of the file, took me a couple of minutes to catch that because the code wasn't working. As I am a newbie in python, I assume that other newbies may have the same issue when trying to run, and then have to troubleshoot, I decided to propose this change so others don't have to 'suffer' :).